flickr30k_experiment/requirements.txt
datasets==3.5.1
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
pycocoevalcap==1.2
sentencepiece==0.2.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
wandb==0.19.10
open_clip_torch==2.21.0